+++ src/lib/lib.ml
@@ -192,6 +192,9 @@ struct
       if len = 0 then [] else String.sub s i len :: loop (i + len)
     in loop 0
 
+  let strip_control_chars s =
+    String.map (fun c -> if c < ' ' then ' ' else c) s
+
   let rec find_from_opt f s i =
     if i = String.length s then
       None