+++ src/lib/lib.mli
@@ -168,6 +168,7 @@ sig
   val chop_suffix : string -> string -> string option
   val lightweight_escaped : string -> string
   val levenshtein_distance : string -> string -> int
+  val strip_control_chars : string -> string
 end
 
 module CRC :